Refactor server api implementation
- Add app related property fields to PbenchServerConfig
- Modularize api implementation
- Reuse loaded pbench server config
- Remove sys.sxit calls from getconfig utility method
